+++ web/components/subscribe-card.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { CopyButton } from "./copy-button";
 
 interface SubscribeCardProps {